Tour.getDistance tested and fixed

+++ src/main/java/tsp/lee/jacobson/Tour.java
@@ -13,7 +13,7 @@ public class Tour{
     private ArrayList<City> tour = new ArrayList<City>();
     // Cache
     private double fitness = 0;
-    private int distance = 0;
+    private double distance = 0;
     
     // Constructs a blank tour
     public Tour(){
@@ -58,9 +58,9 @@ public class Tour{
     }
     
     // Gets the total distance of the tour
-    public int getDistance(){
+    public double getDistance(){
         if (distance == 0) {
-            int tourDistance = 0;
+        	double tourDistance = 0;
             // Loop through our tour's cities
             for (int cityIndex=0; cityIndex < tourSize(); cityIndex++) {
                 // Get city we're travelling from
